@@ -3,56 +3,59 @@ MCP Location Server implementation.
 Generated By: Cursor (Claude Sonnet 4)
 """
 
+from __future__ import annotations
+
 import asyncio
 import logging
-from typing import Any, Dict
+import sys
+from typing import Any
 
+import httpx
 from mcp.server.fastmcp import FastMCP
 from mcp.types import ToolAnnotations
 from pydantic import ValidationError
-import httpx
 
-from .models import GeocodeRequest, GeocodeResponse, ElevationRequest, ElevationResponse
-from .geocoding import get_geocoding_service, close_geocoding_service
-from .elevation import get_elevation_service, close_elevation_service
+from .elevation import close_elevation_service, get_elevation_service
+from .geocoding import close_geocoding_service, get_geocoding_service
+from .models import ElevationRequest, GeocodeRequest
 
 
 class LocationServer:
     """Location MCP Server for geocoding addresses to coordinates."""
-    
+
     def __init__(self) -> None:
         self.logger = logging.getLogger("mcp_location_server.LocationServer")
         self.logger.info("Initializing Location MCP Server")
-    
+
     async def geocode_address(
-        self, 
-        query: str, 
-        limit: int = 1
-    ) -> Dict[str, Any]:
+        self,
+        query: str,
+        limit: int = 1,
+    ) -> dict[str, Any]:
         """
         Geocode an address or location name to latitude/longitude coordinates.
-        
+
         Args:
             query: The address or location to geocode (e.g., "1600 Pennsylvania Avenue, Washington DC")
             limit: Maximum number of results to return (1-10, default: 1)
-            
+
         Returns:
             Dictionary containing geocoding results
         """
-        self.logger.info(f"Geocoding request: query='{query}', limit={limit}")
-        
+        self.logger.info("Geocoding request: query='%s', limit=%d", query, limit)
+
         try:
             # Validate the request
             request = GeocodeRequest(query=query, limit=limit)
-            
+
             # Get the geocoding service
             service = await get_geocoding_service()
-            
+
             # Perform geocoding
             response = await service.geocode(request)
-            
-            self.logger.info(f"Geocoding successful: found {response.count} results")
-            
+
+            self.logger.info("Geocoding successful: found %d results", response.count)
+
             # Convert to dictionary for MCP response
             return {
                 "results": [
@@ -71,27 +74,27 @@ class LocationServer:
                 "query": response.query,
                 "count": response.count,
             }
-            
-        except ValidationError as e:
-            self.logger.error(f"Request validation error: {e}")
+
+        except ValidationError:
+            self.logger.exception("Request validation error")
             return {
                 "error": "Invalid request",
-                "message": str(e),
+                "message": "Invalid request parameters",
                 "results": [],
                 "query": query,
                 "count": 0,
             }
-        except httpx.HTTPError as e:
-            self.logger.error(f"HTTP error during geocoding: {e}")
+        except httpx.HTTPError:
+            self.logger.exception("HTTP error during geocoding")
             return {
                 "error": "API request failed",
-                "message": str(e),
+                "message": "Failed to connect to geocoding service",
                 "results": [],
                 "query": query,
                 "count": 0,
             }
-        except Exception as e:
-            self.logger.error(f"Unexpected error during geocoding: {e}")
+        except Exception:
+            self.logger.exception("Unexpected error during geocoding")
             return {
                 "error": "Internal server error",
                 "message": "An unexpected error occurred",
@@ -99,31 +102,31 @@ class LocationServer:
                 "query": query,
                 "count": 0,
             }
-    
+
     async def reverse_geocode_coordinates(
-        self, 
-        latitude: float, 
-        longitude: float
-    ) -> Dict[str, Any]:
+        self,
+        latitude: float,
+        longitude: float,
+    ) -> dict[str, Any]:
         """
         Reverse geocode coordinates to an address.
-        
+
         Args:
             latitude: Latitude in decimal degrees
             longitude: Longitude in decimal degrees
-            
+
         Returns:
             Dictionary containing reverse geocoding result
         """
-        self.logger.info(f"Reverse geocoding request: lat={latitude}, lon={longitude}")
-        
+        self.logger.info("Reverse geocoding request: lat=%f, lon=%f", latitude, longitude)
+
         try:
             # Get the geocoding service
             service = await get_geocoding_service()
-            
+
             # Perform reverse geocoding
             result = await service.reverse_geocode(latitude, longitude)
-            
+
             if result is None:
                 self.logger.info("Reverse geocoding: no result found")
                 return {
@@ -132,9 +135,7 @@ class LocationServer:
                     "longitude": longitude,
                     "found": False,
                 }
-            
             self.logger.info("Reverse geocoding successful")
-            
             return {
                 "result": {
                     "display_name": result.display_name,
@@ -150,29 +151,29 @@ class LocationServer:
                 "longitude": longitude,
                 "found": True,
             }
-            
-        except ValueError as e:
-            self.logger.error(f"Invalid coordinates for reverse geocoding: {e}")
+
+        except ValueError:
+            self.logger.exception("Invalid coordinates for reverse geocoding")
             return {
                 "error": "Invalid coordinates",
-                "message": str(e),
+                "message": "Coordinates are out of valid range",
                 "result": None,
                 "latitude": latitude,
                 "longitude": longitude,
                 "found": False,
             }
-        except httpx.HTTPError as e:
-            self.logger.error(f"HTTP error during reverse geocoding: {e}")
+        except httpx.HTTPError:
+            self.logger.exception("HTTP error during reverse geocoding")
             return {
                 "error": "API request failed",
-                "message": str(e),
+                "message": "Failed to connect to geocoding service",
                 "result": None,
                 "latitude": latitude,
                 "longitude": longitude,
                 "found": False,
             }
-        except Exception as e:
-            self.logger.error(f"Unexpected error during reverse geocoding: {e}")
+        except Exception:
+            self.logger.exception("Unexpected error during reverse geocoding")
             return {
                 "error": "Internal server error",
                 "message": "An unexpected error occurred",
@@ -181,97 +182,96 @@ class LocationServer:
                 "longitude": longitude,
                 "found": False,
             }
-    
+
     async def get_elevation(
         self,
         latitude: float,
         longitude: float,
-        dataset: str = "srtm90m"
-    ) -> Dict[str, Any]:
+        dataset: str = "srtm90m",
+    ) -> dict[str, Any]:
         """
         Get elevation data for coordinates.
-        
+
         Args:
             latitude: Latitude in decimal degrees (-90 to 90)
             longitude: Longitude in decimal degrees (-180 to 180)
             dataset: Elevation dataset to use (default: srtm90m)
-            
+
         Returns:
             Dictionary containing elevation data in meters and feet
         """
-        self.logger.info(f"Elevation request: lat={latitude}, lon={longitude}, dataset={dataset}")
-        
+        self.logger.info("Elevation request: lat=%f, lon=%f, dataset=%s", latitude, longitude, dataset)
+
         try:
             # Validate the request
             request = ElevationRequest(
                 latitude=latitude,
                 longitude=longitude,
-                dataset=dataset
+                dataset=dataset,
             )
-            
+
             # Get the elevation service
             service = await get_elevation_service()
-            
+
             # Get elevation data
             response = await service.get_elevation(request)
-            
+
             if response.results:
                 result = response.results[0]
-                self.logger.info(f"Elevation successful: {result.elevation.meters}m")
-                
+                self.logger.info("Elevation successful: %fm", result.elevation.meters)
+
                 return {
                     "latitude": result.latitude,
                     "longitude": result.longitude,
                     "elevation_meters": result.elevation.meters,
                     "elevation_feet": result.elevation.feet,
                     "dataset": result.elevation.dataset,
-                    "found": True
+                    "found": True,
                 }
-            else:
-                self.logger.info("Elevation: no data available for coordinates")
-                return {
-                    "latitude": latitude,
-                    "longitude": longitude,
-                    "elevation_meters": None,
-                    "elevation_feet": None,
-                    "dataset": dataset,
-                    "found": False,
-                    "message": "No elevation data available for these coordinates"
-                }
-            
-        except ValidationError as e:
-            self.logger.error(f"Request validation error: {e}")
+            self.logger.info("Elevation: no data available for coordinates")
+            return {
+                "latitude": latitude,
+                "longitude": longitude,
+                "elevation_meters": None,
+                "elevation_feet": None,
+                "dataset": dataset,
+                "found": False,
+                "message": "No elevation data available for these coordinates",
+            }
+
+        except ValidationError:
+            self.logger.exception("Request validation error")
             return {
                 "error": "Invalid request",
-                "message": str(e),
+                "message": "Invalid request parameters",
                 "latitude": latitude,
                 "longitude": longitude,
-                "found": False
+                "found": False,
             }
-        except httpx.HTTPError as e:
-            self.logger.error(f"HTTP error during elevation lookup: {e}")
+        except httpx.HTTPError:
+            self.logger.exception("HTTP error during elevation lookup")
             return {
-                "error": "API request failed", 
-                "message": str(e),
+                "error": "API request failed",
+                "message": "Failed to connect to elevation service",
                 "latitude": latitude,
                 "longitude": longitude,
-                "found": False
+                "found": False,
             }
-        except Exception as e:
-            self.logger.error(f"Unexpected error during elevation lookup: {e}")
+        except Exception:
+            self.logger.exception("Unexpected error during elevation lookup")
             return {
                 "error": "Internal server error",
                 "message": "An unexpected error occurred",
                 "latitude": latitude,
                 "longitude": longitude,
-                "found": False
+                "found": False,
             }
 
 
 async def create_server() -> FastMCP:
     """Create and configure the MCP server."""
     logging.basicConfig(level=logging.INFO)
-    
+
     # Create the MCP server
     mcp = FastMCP(
         "location_mcp_server",
@@ -285,10 +285,10 @@ async def create_server() -> FastMCP:
             "You can handle addresses, cities, landmarks, and points of interest worldwide."
         ),
     )
-    
+
     # Create the location server instance
     location_server = LocationServer()
-    
+
     # Add the geocoding tool
     mcp.add_tool(
         location_server.geocode_address,
@@ -309,7 +309,7 @@ async def create_server() -> FastMCP:
         ),
         structured_output=True,
     )
-    
+
     # Add the reverse geocoding tool
     mcp.add_tool(
         location_server.reverse_geocode_coordinates,
@@ -329,11 +329,11 @@ async def create_server() -> FastMCP:
         ),
         structured_output=True,
     )
-    
+
     # Add the elevation tool
     mcp.add_tool(
         location_server.get_elevation,
-        "get_elevation", 
+        "get_elevation",
         title="Get elevation data for coordinates",
         description=(
             "Get elevation data for latitude/longitude coordinates. "
@@ -351,12 +351,12 @@ async def create_server() -> FastMCP:
         ),
         structured_output=True,
     )
-    
+
     return mcp
 
 
 # Global app instance for the entry point
-app: FastMCP = None
+app: FastMCP | None = None
 
 
 async def get_app() -> FastMCP:
@@ -369,23 +369,23 @@ async def get_app() -> FastMCP:
 
 async def main() -> None:
     """Main entry point for running the server."""
-    import sys
-    
+    logger = logging.getLogger(__name__)
+
     try:
         server = await create_server()
-        
+
         # Check if running in STDIO mode (for MCP clients like Cursor)
         # or HTTP mode (for testing/debugging)
         if len(sys.argv) > 1 and sys.argv[1] == "--http":
-            logging.info("Starting server in HTTP mode")
+            logger.info("Starting server in HTTP mode")
             await server.run_streamable_http_async()
         else:
-            logging.info("Starting server in STDIO mode for MCP client")
+            logger.info("Starting server in STDIO mode for MCP client")
             await server.run_stdio_async()
     except KeyboardInterrupt:
-        logging.info("Server stopped by user")
-    except Exception as e:
-        logging.error(f"Server error: {e}")
+        logger.info("Server stopped by user")
+    except Exception:
+        logger.exception("Server error")
         raise
     finally:
         # Clean up the services
@@ -394,4 +394,4 @@ async def main() -> None:
 
 
 if __name__ == "__main__":
-    asyncio.run(main()) 
+    asyncio.run(main())
