@@ -1,1 +1,1 @@
-# Generated By: Cursor (Claude Sonnet 4) 
+# Generated By: Cursor (Claude Sonnet 4)
