@@ -3,12 +3,19 @@ Tests for the geocoding functionality.
 Generated By: Cursor (Claude Sonnet 4)
 """
 
-import pytest
-from unittest.mock import AsyncMock, MagicMock, patch
-import httpx
+from unittest.mock import MagicMock, patch
 
-from mcp_location_server.models import GeocodeRequest, GeocodeResult, GeocodeResponse
+import httpx
+import pytest
+
 from mcp_location_server.geocoding import NominatimGeocodingService
+from mcp_location_server.models import GeocodeRequest, GeocodeResponse, GeocodeResult
+
+# Test constants
+TEST_DC_LATITUDE = 38.8976633
+TEST_DC_LONGITUDE = -77.0365739
+TEST_IMPORTANCE = 0.9
+TEST_LIMIT = 5
 
 
 class TestNominatimGeocodingService:
@@ -27,11 +34,11 @@ class TestNominatimGeocodingService:
                 "type": "house",
                 "class": "place",
                 "osm_id": 123456,
-                "osm_type": "way"
-            }
+                "osm_type": "way",
+            },
         ]
 
-        with patch.object(httpx.AsyncClient, 'get') as mock_get:
+        with patch.object(httpx.AsyncClient, "get") as mock_get:
             # Setup mock response
             mock_response = MagicMock()
             mock_response.raise_for_status = MagicMock()
@@ -51,16 +58,16 @@ class TestNominatimGeocodingService:
             assert len(result.results) == 1
 
             location = result.results[0]
-            assert location.latitude == 38.8976633
-            assert location.longitude == -77.0365739
-            assert location.importance == 0.9
+            assert location.latitude == TEST_DC_LATITUDE
+            assert location.longitude == TEST_DC_LONGITUDE
+            assert location.importance == TEST_IMPORTANCE
             assert location.place_type == "house"
             assert location.place_class == "place"
 
     @pytest.mark.asyncio
     async def test_geocode_empty_results(self):
         """Test geocoding with no results found."""
-        with patch.object(httpx.AsyncClient, 'get') as mock_get:
+        with patch.object(httpx.AsyncClient, "get") as mock_get:
             # Setup mock response with empty results
             mock_response = MagicMock()
             mock_response.raise_for_status = MagicMock()
@@ -81,7 +88,7 @@ class TestNominatimGeocodingService:
     @pytest.mark.asyncio
     async def test_geocode_http_error(self):
         """Test geocoding with HTTP error."""
-        with patch.object(httpx.AsyncClient, 'get') as mock_get:
+        with patch.object(httpx.AsyncClient, "get") as mock_get:
             # Setup mock to raise HTTP error
             mock_get.side_effect = httpx.HTTPError("API request failed")
 
@@ -103,10 +110,10 @@ class TestNominatimGeocodingService:
             "type": "house",
             "class": "place",
             "osm_id": 123456,
-            "osm_type": "way"
+            "osm_type": "way",
         }
 
-        with patch.object(httpx.AsyncClient, 'get') as mock_get:
+        with patch.object(httpx.AsyncClient, "get") as mock_get:
             mock_response = MagicMock()
             mock_response.raise_for_status = MagicMock()
             mock_response.json = MagicMock(return_value=mock_response_data)
@@ -119,14 +126,14 @@ class TestNominatimGeocodingService:
             # Assertions
             assert result is not None
             assert isinstance(result, GeocodeResult)
-            assert result.latitude == 38.8976633
-            assert result.longitude == -77.0365739
+            assert result.latitude == TEST_DC_LATITUDE
+            assert result.longitude == TEST_DC_LONGITUDE
             assert result.place_type == "house"
 
     @pytest.mark.asyncio
     async def test_reverse_geocode_no_result(self):
         """Test reverse geocoding with no result."""
-        with patch.object(httpx.AsyncClient, 'get') as mock_get:
+        with patch.object(httpx.AsyncClient, "get") as mock_get:
             mock_response = MagicMock()
             mock_response.raise_for_status = MagicMock()
             mock_response.json = MagicMock(return_value={"error": "Unable to geocode"})
@@ -153,7 +160,7 @@ class TestNominatimGeocodingService:
         """Test async context manager functionality."""
         async with NominatimGeocodingService() as service:
             assert service is not None
-            assert hasattr(service, 'client')
+            assert hasattr(service, "client")
 
         # After exiting context, client should be closed
         # Note: In real usage, the client would be closed
@@ -194,7 +201,7 @@ class TestGeocodeRequest:
         """Test creating a valid geocode request."""
         request = GeocodeRequest(query="New York", limit=5)
         assert request.query == "New York"
-        assert request.limit == 5
+        assert request.limit == TEST_LIMIT
 
     def test_default_limit(self):
         """Test default limit value."""
@@ -203,21 +210,21 @@ class TestGeocodeRequest:
 
     def test_invalid_limit_too_low(self):
         """Test validation for limit too low."""
-        with pytest.raises(ValueError):
+        with pytest.raises(ValueError, match="limit"):
             GeocodeRequest(query="London", limit=0)
 
     def test_invalid_limit_too_high(self):
         """Test validation for limit too high."""
-        with pytest.raises(ValueError):
+        with pytest.raises(ValueError, match="limit"):
             GeocodeRequest(query="London", limit=20)
 
     def test_empty_query(self):
         """Test validation for empty query."""
-        with pytest.raises(ValueError):
+        with pytest.raises(ValueError, match="query"):
             GeocodeRequest(query="")
 
     def test_query_too_long(self):
         """Test validation for query too long."""
         long_query = "x" * 501  # Exceeds max length of 500
-        with pytest.raises(ValueError):
+        with pytest.raises(ValueError, match="query"):
             GeocodeRequest(query=long_query)
